feature(CP-59): Added Update User, email exists validation and update Test

+++ mp/src/main/java/ro/ubb/mp/service/user/UserServiceImpl.java
@@ -2,8 +2,6 @@ package ro.ubb.mp.service.user;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import ro.ubb.mp.controller.dto.UserRequestDTO;
 import ro.ubb.mp.dao.model.Role;
@@ -12,7 +10,6 @@ import ro.ubb.mp.dao.repository.RoleRepository;
 import ro.ubb.mp.dao.repository.UserRepository;
 
 import javax.transaction.Transactional;
-import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 
@@ -36,20 +33,17 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public Optional<User> getUserById(Long id) {
-        //log.info("fetching user {}", userRepository.findById(id).get().getFullName());
-
         return userRepository.findById(id);
     }
 
     @Override
     public User getUserByEmail(String email) {
-        //log.info("fetching user {}", userRepository.findByEmail(email).getFullName());
-
         return userRepository.findByEmail(email);
     }
 
     @Override
     public User saveUser(UserRequestDTO userDTO) {
+        validateEmailExists(userDTO.getEmail());
         final User userToBeSaved = User.builder()
                 .fullName(userDTO.getFullName())
                 .email(userDTO.getEmail())
@@ -59,21 +53,21 @@ public class UserServiceImpl implements UserService {
                 .interestAreas(userDTO.getInterestAreas())
                 .build();
         userToBeSaved.getRoles().add(userDTO.getRole());
-        //log.info("saving new user {} to the database", userToBeSaved.getFullName());
         return userRepository.save(userToBeSaved);
     }
 
     @Override
     public Role saveRole(Role role) {
-        //log.info("saving new role {} to the database", role.getName());
-
         return roleRepository.save(role);
     }
 
     @Override
-    public void addRoleToUser(String email, String roleName) {
-        //log.info("adding role {} to user {}", roleName, userRepository.findByEmail(email).getFullName());
+    public void deleteRole(Role role) {
+        roleRepository.deleteById(role.getId());
+    }
 
+    @Override
+    public void addRoleToUser(String email, String roleName) {
         User user = userRepository.findByEmail((email));
         Role role = roleRepository.findByName(roleName);
         user.getRoles().add(role);
@@ -81,11 +75,45 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUserByEmail(String email) {
-        userRepository.deleteById(getUserByEmail(email).getId());
+    public User updateUser(Long id, UserRequestDTO userDTO) {
+        User userToUpdate = userRepository.findById(id).orElseThrow(
+                () -> new IllegalStateException(String.format("User with id %s doesn't exist", id))
+        );
+        validateEmailExists(userDTO.getEmail());
+        userToUpdate.setFullName(userDTO.getFullName());
+        userToUpdate.setEmail(userDTO.getEmail());
+        userToUpdate.setPassword(userDTO.getPassword());
+        userToUpdate.setInterestAreas(userDTO.getInterestAreas());
+        userToUpdate.setProfilePicture(userDTO.getProfilePic());
+        userToUpdate.getRoles().add(userDTO.getRole());
+
+        return userRepository.save(userToUpdate);
     }
+
+    @Override
+    public void deleteUserByEmail(String email) {
+        boolean userExists = userRepository.existsById(userRepository.findByEmail(email).getId());
+        if (userExists) {
+            userRepository.deleteById(getUserByEmail(email).getId());
+        } else {
+            throw new IllegalStateException(String.format("User with email %s doesn't exist", email));
+        }
+    }
+
     @Override
     public void deleteUserById(Long id) {
-        userRepository.deleteById(id);
+        boolean userExists = userRepository.existsById(id);
+        if (userExists) {
+            userRepository.deleteById(id);
+        } else {
+            throw new IllegalStateException(String.format("User with id %s doesn't exist", id));
+        }
+    }
+
+    void validateEmailExists(String email) {
+        Optional<User> userOptional = Optional.ofNullable(userRepository.findByEmail(email));
+        if (userOptional.isPresent()) {
+            throw new IllegalStateException(String.format("Email address %s already exists", email));
+        }
     }
 }
